% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumiR.R
\name{s_agrupada}
\alias{s_agrupada}
\title{Medidas de resumen univariadas y Tablas de Frecuencias para datos numéricos discretos o continuos}
\usage{
s_agrupada(
  x,
  decimales = 1,
  li = NULL,
  ls = NULL,
  a = NULL,
  derecha = TRUE,
  imprimir = TRUE,
  clipboard = FALSE,
  grafico = "ninguno",
  frec = "relativa",
  pf = FALSE,
  ...
)
}
\arguments{
\item{x}{vector numérico de tipo *integer* o *numeric*; en este último caso, el vector debe tener una longitud mayor a 1, y permitir el cálculo de medidas de variabilidad.}

\item{decimales}{el número de posiciones decimales de los resultados.}

\item{li}{límite inferior del primer intervalo de clase, para datos de tipo *numeric*.}

\item{ls}{límite superior del intervalo de clase superior, para datos de tipo *numeric*.}

\item{a}{amplitud del intervalo de clase, para datos de tipo *numeric*.}

\item{derecha}{si los intervalos de clase son cerrados por derecha, para datos de tipo *numeric*.}

\item{imprimir}{lógico; si es verdadero (TRUE), imprime los resultados en la pantalla de la consola; si es falso (FALSE), devuelve un data.frame; por defecto es TRUE.}

\item{clipboard}{lógico; es es verdadero (TRUE), exporta los resultados con el portapapeles; por defecto es FALSE}

\item{grafico}{texto indicando el tipo de gráfico a realizar: frecuencias simples, acumuladas o ambas; los valores posibles son "ninguno", "fs" y "fa"; por defecto es "ninguno"; hereda propiedades de ggplot()}

\item{frec}{texto indicando tipo de frecuencia a utilizar en caso de realizar un gráfico; los valores posibles son "absoluta" y "relativa"; por defecto es "absoluta".}

\item{pf}{lógico; si es verdadero (TRUE), devuelve el polígono de frecuencias en lugar del histograma de frecuencias, cuando x es continua; por defecto es FALSE.}

\item{...}{argumentos que heredan de la función ggplot()}
}
\value{
cuando el argumento imprimir = T, imprime una tabla de frecuencias para datos continuos o discretos; cuando el argumento gráfico es "fs" o "fa" devuelve el gráfico de frecuencias simples y acumulados respeectivamente, adecuados para cada tipo de variable; cuando el argumento imprimir=FALSE, devuelve una lista.
}
\description{
Devuelve la tabla de frecuencias de una serie de datos numérica, univariada, de tipo *integer* o *numeric*. En el caso de datos de tipo *numeric*, es posible especificar los límites y la amplitud de los intervalos de clase. Por defecto, construye los intervalos de clase cerrados por derecha, con la regla de Sturges.
Si se desea, es posible exportar los resultados en el portapapeles.
}
\examples{
library(resumiR)
x = iris[[1]]
s_agrupada(x, li=4, ls=8, a=1, clipboard=TRUE, grafico="fs")
s_agrupada(x, imprimir=FALSE, grafico="fs",pf=TRUE)
s_agrupada(x, grafico="fa")
}
